Move some components of geocatalog
