+++ bundles/dbjobs/src/main/java/org/orbisgis/dbjobs/service/DriverFunctionContainerImpl.java
@@ -54,6 +54,7 @@ import java.io.File;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.concurrent.ExecutorService;
@@ -98,6 +99,11 @@ public class DriverFunctionContainerImpl implements DriverFunctionContainer {
     @Override
     public void removeDriverFunction(DriverFunction driverFunction) {
         fileDrivers.remove(driverFunction);
+    }
+
+    @Override
+    public List<DriverFunction> getDriverFunctionList() {
+        return Collections.unmodifiableList(fileDrivers);
     }
 
     @Override
